+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/controller/ConfiguredApiController.kt
@@ -22,8 +22,8 @@ import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.Configu
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.ConfiguredApiController.FiltersPrefix.FILTERS_QUERY_DESCRIPTION
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.ConfiguredApiController.FiltersPrefix.FILTERS_QUERY_EXAMPLE
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.Count
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementExecutionResponse
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementExecutionStatus
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementResult
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.exception.NoDataAvailableException
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.security.DprAuthAwareAuthenticationToken
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.ConfiguredApiService
@@ -117,8 +117,9 @@ class ConfiguredApiController(val configuredApiService: ConfiguredApiService) {
 
   @GetMapping("/async/reports/{reportId}/{reportVariantId}")
   @Operation(
-    description = "Executes asynchronously the dataset query for the given report and returns the execution ID. " +
-      "This is the asynchronous non-paginated version of the /reports/{reportId}/{reportVariantId} API.",
+    description = "Executes asynchronously the dataset query for the given report and stores the result into an external table." +
+      "The response returned contains the table ID and the execution ID. " +
+      "This is the asynchronous version of the /reports/{reportId}/{reportVariantId} API.",
     security = [ SecurityRequirement(name = "bearer-jwt") ],
     responses = [
       ApiResponse(
@@ -149,7 +150,7 @@ class ConfiguredApiController(val configuredApiService: ConfiguredApiService) {
     @RequestParam("dataProductDefinitionsPath", defaultValue = ReportDefinitionController.DATA_PRODUCT_DEFINITIONS_PATH_EXAMPLE)
     dataProductDefinitionsPath: String? = null,
     authentication: Authentication,
-  ): ResponseEntity<String> {
+  ): ResponseEntity<StatementExecutionResponse> {
     return try {
       ResponseEntity
         .status(HttpStatus.OK)
@@ -222,9 +223,10 @@ class ConfiguredApiController(val configuredApiService: ConfiguredApiService) {
     }
   }
 
-  @GetMapping("/reports/{reportId}/{reportVariantId}/statements/{statementId}/result")
+  @GetMapping("/reports/{reportId}/{reportVariantId}/tables/{tableId}/result")
   @Operation(
-    description = "Returns the resulting rows of the executed statement.",
+    description = "Returns the resulting rows of the executed statement in a paginated " +
+      "fashion which has been have been stored in a dedicated table.",
     security = [ SecurityRequirement(name = "bearer-jwt") ],
     responses = [
       ApiResponse(
@@ -242,23 +244,27 @@ class ConfiguredApiController(val configuredApiService: ConfiguredApiService) {
     @PathVariable("reportVariantId") reportVariantId: String,
     @RequestParam("dataProductDefinitionsPath", defaultValue = ReportDefinitionController.DATA_PRODUCT_DEFINITIONS_PATH_EXAMPLE)
     dataProductDefinitionsPath: String? = null,
-    @PathVariable("statementId") statementId: String,
-    @Parameter(
-      description = "A value that indicates the starting point for the next set of response records in a subsequent request. " +
-        "If a value is returned in a response, you can retrieve the next set of records by providing this returned NextToken value in " +
-        "the next nextToken query parameter and retrying the API call. " +
-        "If the nextToken field is empty, all response records have been retrieved for the request.",
-      example = "aa37926e-e40a-43ce-a553-ec015ecec52e",
-    )
-    @RequestParam("nextToken", required = false)
-    nextToken: String?,
+    @PathVariable("tableId") tableId: String,
+    @RequestParam(defaultValue = "1")
+    @Min(1)
+    selectedPage: Long,
+    @RequestParam(defaultValue = "10")
+    @Min(1)
+    pageSize: Long,
     authentication: Authentication,
-  ): ResponseEntity<StatementResult> {
+  ): ResponseEntity<List<Map<String, Any?>>> {
     return try {
       ResponseEntity
         .status(HttpStatus.OK)
         .body(
-          configuredApiService.getStatementResult(statementId, reportId, reportVariantId, dataProductDefinitionsPath, nextToken),
+          configuredApiService.getStatementResult(
+            tableId,
+            reportId,
+            reportVariantId,
+            dataProductDefinitionsPath,
+            selectedPage,
+            pageSize,
+          ),
         )
     } catch (exception: NoDataAvailableException) {
       val headers = HttpHeaders()