+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/RedshiftDataApiRepository.kt
@@ -1,6 +1,10 @@
 package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data
 
+import org.apache.commons.lang3.time.StopWatch
 import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate
 import org.springframework.stereotype.Service
 import software.amazon.awssdk.services.redshiftdata.RedshiftDataClient
 import software.amazon.awssdk.services.redshiftdata.model.ColumnMetadata
@@ -8,11 +12,11 @@ import software.amazon.awssdk.services.redshiftdata.model.DescribeStatementReque
 import software.amazon.awssdk.services.redshiftdata.model.ExecuteStatementRequest
 import software.amazon.awssdk.services.redshiftdata.model.ExecuteStatementResponse
 import software.amazon.awssdk.services.redshiftdata.model.Field
-import software.amazon.awssdk.services.redshiftdata.model.GetStatementResultRequest
 import software.amazon.awssdk.services.redshiftdata.model.GetStatementResultResponse
 import software.amazon.awssdk.services.redshiftdata.model.SqlParameter
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementExecutionResponse
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementExecutionStatus
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementResult
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.TableIdGenerator
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
@@ -20,6 +24,9 @@ import java.time.format.DateTimeFormatter
 class RedshiftDataApiRepository(
   val redshiftDataClient: RedshiftDataClient,
   val executeStatementRequestBuilder: ExecuteStatementRequest.Builder,
+  val tableIdGenerator: TableIdGenerator,
+  @Value("\${dpr.lib.redshiftdataapi.s3location:#{'dpr-working-development/reports'}}")
+  private val s3location: String = "dpr-working-development/reports",
 ) : RepositoryHelper() {
 
   companion object {
@@ -34,15 +41,26 @@ class RedshiftDataApiRepository(
     policyEngineResult: String,
     dynamicFilterFieldId: String? = null,
     dataSourceName: String,
-  ): String {
+  ): StatementExecutionResponse {
+    val tableId = tableIdGenerator.generateNewExternalTableId()
+    val generateSql = """
+          CREATE EXTERNAL TABLE reports.$tableId 
+          STORED AS parquet 
+          LOCATION 's3://$s3location/$tableId/' 
+          AS ( 
+          ${
+      buildFinalQuery(
+        buildReportQuery(query),
+        buildPolicyQuery(policyEngineResult),
+        buildFiltersQuery(filters, queryParamKeyTransformer),
+        buildFinalStageQuery(dynamicFilterFieldId, sortColumn, sortedAsc),
+      )
+    }
+          );
+    """.trimIndent()
     val requestBuilder = executeStatementRequestBuilder
       .sql(
-        buildFinalQuery(
-          buildReportQuery(query),
-          buildPolicyQuery(policyEngineResult),
-          buildFiltersQuery(filters, queryParamKeyTransformer),
-          buildFinalStageQuery(dynamicFilterFieldId, sortColumn, sortedAsc),
-        ),
+        generateSql,
       )
     if (filters.isNotEmpty()) {
       requestBuilder
@@ -51,8 +69,9 @@ class RedshiftDataApiRepository(
     val statementRequest: ExecuteStatementRequest = requestBuilder.build()
 
     val response: ExecuteStatementResponse = redshiftDataClient.executeStatement(statementRequest)
-    log.info("Execution ID: {}", response.id())
-    return response.id()
+    log.debug("Execution ID: {}", response.id())
+    log.debug("External table ID: {}", tableId)
+    return StatementExecutionResponse(tableId, response.id())
   }
 
   fun getStatementStatus(statementId: String): StatementExecutionStatus {
@@ -70,14 +89,24 @@ class RedshiftDataApiRepository(
     )
   }
 
-  fun getStatementResult(statementId: String, nextToken: String? = null): StatementResult {
-    val requestBuilder = GetStatementResultRequest.builder()
-    requestBuilder.id(statementId)
-    nextToken?.let { requestBuilder.nextToken(it) }
-    val statementRequest: GetStatementResultRequest = requestBuilder
-      .build()
-    val resultStatementResponse: GetStatementResultResponse = redshiftDataClient.getStatementResult(statementRequest)
-    return StatementResult(extractRecords(resultStatementResponse), resultStatementResponse.nextToken())
+  fun getPaginatedExternalTableResult(
+    tableId: String,
+    selectedPage: Long,
+    pageSize: Long,
+    jdbcTemplate: NamedParameterJdbcTemplate = populateJdbcTemplate(),
+  ): List<Map<String, Any?>> {
+    val stopwatch = StopWatch.createStarted()
+    val result = jdbcTemplate
+      .queryForList(
+        "SELECT * FROM reports.$tableId limit $pageSize OFFSET ($selectedPage - 1) * $pageSize;",
+        MapSqlParameterSource(),
+      )
+      .map {
+        transformTimestampToLocalDateTime(it)
+      }
+    stopwatch.stop()
+    log.debug("Query Execution time in ms: {}", stopwatch.time)
+    return result
   }
 
   private fun extractRecords(resultStatementResponse: GetStatementResultResponse) =