+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/model/redshiftdata/StatementExecutionResponse.kt
@@ -1,0 +1,6 @@
+package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata
+
+data class StatementExecutionResponse(
+  val tableId: String,
+  val executionId: String,
+)