+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/service/TableIdGenerator.kt
@@ -1,0 +1,12 @@
+package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service
+
+import org.springframework.stereotype.Service
+import java.util.UUID
+
+@Service
+class TableIdGenerator {
+
+  fun generateNewExternalTableId(): String {
+    return "_" + UUID.randomUUID().toString().replace("-", "_")
+  }
+}